Crate common_primitives

Frequency Primitives

Primitives package contains many of the structs and trait implementations for Pallets and utilities that need to be shared across packages

Modules

messages

Structs and traits for the Messages pallet.

msa

Structs and traits for the MSA pallet.

node

Node level primitives.

parquet

Structs and traits for parquet

schema

Structs and traits for the Schema pallet

utils

Structs and traits for the utility package.