Module common_primitives::msa

Structs and traits for the MSA pallet.

Re-exports

pub use crate::schema::SchemaId;

Structs

Delegation

Struct for the information of the relationship between an MSA and a Provider

DelegatorId

A DelegatorId an MSA Id serving the role of a Delegator. Delegators delegate to Providers. Encodes and Decodes as just a u64

KeyInfoResponse

RPC Response for getting getting MSA ids

ProviderId

Provider is the recipient of a delegation. It is a subset of an MSA Encodes and Decodes as just a u64

ProviderRegistryEntry

This is the metadata associated with a provider. As of now it is just a name, but it will likely be expanded in the future

Constants

EXPECTED_MAX_NUMBER_OF_PROVIDERS_PER_DELEGATOR

Maximum # of providers per delegator for deletion/benchmarking use

Traits

DelegationValidator

A behavior that allows for validating a delegator-provider relationship

MsaLookup

A behavior that allows looking up an MSA id

MsaValidator

A behavior that allows for validating an MSA

ProviderLookup

A behavior that allows for looking up delegator-provider relationships

SchemaGrantValidator

A behavior that allows for validating a schema grant

Type Definitions

MessageSourceId

Message Source Id or msaId is the unique identifier for Message Source Accounts