Trait common_primitives::msa::AccountProvider

pub trait AccountProvider {
    type AccountId;
    type BlockNumber;
    type MaxSchemaGrants: Get<u32> + Clone + Eq;

    fn get_msa_id(key: &Self::AccountId) -> Option<MessageSourceId>;
    fn get_provider_info_of(
        delegator: Delegator,
        provider: Provider
    ) -> Option<ProviderInfo<Self::BlockNumber, Self::MaxSchemaGrants>>;
    fn ensure_valid_msa_key(
        key: &Self::AccountId
    ) -> Result<MessageSourceId, DispatchError>;
    fn ensure_valid_delegation(
        provider: Provider,
        delegator: Delegator,
        block_number: Option<Self::BlockNumber>
    ) -> Result<ProviderInfo<Self::BlockNumber, Self::MaxSchemaGrants>, DispatchError>;
    fn ensure_valid_schema_grant(
        provider: Provider,
        delegator: Delegator,
        schema_id: SchemaId
    ) -> DispatchResult;
}

This allows other pallets to resolve MSA information.

Required Associated Types

type AccountId

Type used to associate a key to an MSA.

type BlockNumber

Type for block number.

type MaxSchemaGrants: Get<u32> + Clone + Eq

Type for maximum number of schemas that can be granted to a provider.

Required Methods

fn get_msa_id(key: &Self::AccountId) -> Option<MessageSourceId>

Gets the MSA Id associated with this AccountId if any

Arguments

key - The AccountId to lookup

Returns

Option<MessageSourceId>

fn get_provider_info_of(
delegator: Delegator,
provider: Provider
) -> Option<ProviderInfo<Self::BlockNumber, Self::MaxSchemaGrants>>

Gets the relationship information for this delegator, provider pair

Arguments

delegator - The MessageSourceId that delegated to the provider

provider - The MessageSourceId that has been delegated to

Returns

Option<ProviderInfo<Self::BlockNumber>>

fn ensure_valid_msa_key(
key: &Self::AccountId
) -> Result<MessageSourceId, DispatchError>

Check that a key is associated to an MSA and returns key information. Returns a [DispatchError] if there is no MSA associated with the key

Arguments

key - The AccountId to lookup

Returns

Result<MessageSourceId, DispatchError>

fn ensure_valid_delegation(
provider: Provider,
delegator: Delegator,
block_number: Option<Self::BlockNumber>
) -> Result<ProviderInfo<Self::BlockNumber, Self::MaxSchemaGrants>, DispatchError>

Validates that the delegator and provider have a relationship at this point

Arguments

provider - The MessageSourceId that has been delegated to

delegator - The MessageSourceId that delegated to the provider

Returns

DispatchResult The return type of a Dispatchable in frame.

fn ensure_valid_schema_grant(
provider: Provider,
delegator: Delegator,
schema_id: SchemaId
) -> DispatchResult

Validates if the provider is allowed to use the schema

Arguments

provider - The MessageSourceId that has been delegated to

delegator - The MessageSourceId that delegated to the provider

schema_id - The SchemaId that the provider is trying to use

Returns

DispatchResult The return type of a Dispatchable in frame.