Trait common_primitives::msa::ProviderLookup

pub trait ProviderLookup {
    type BlockNumber;
    type MaxSchemaGrantsPerDelegation: Get<u32> + Clone + Eq;
    type SchemaId;

    fn get_delegation_of(
        delegator: DelegatorId,
        provider: ProviderId
    ) -> Option<Delegation<Self::SchemaId, Self::BlockNumber, Self::MaxSchemaGrantsPerDelegation>>;
}

A behavior that allows for looking up delegator-provider relationships

Required Associated Types

type BlockNumber

Type for block number.

type MaxSchemaGrantsPerDelegation: Get<u32> + Clone + Eq

Type for maximum number of schemas that can be granted to a provider.

type SchemaId

Schema Id is the unique identifier for a Schema

Required Methods

fn get_delegation_of(
delegator: DelegatorId,
provider: ProviderId
) -> Option<Delegation<Self::SchemaId, Self::BlockNumber, Self::MaxSchemaGrantsPerDelegation>>

Gets the relationship information for this delegator, provider pair