Module common_primitives::node

Node level primitives.

Modules

generic

Generic implementations of Extrinsic/Header/Block.

Structs

BlakeTwo256

Blake2-256 Hash implementation.

OpaqueExtrinsic

Simple blob to hold an extrinsic without committing to its format and ensure it is serialized correctly.

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

MultiSignature

Signature verify that can work with any known signature types..

Traits

IdentifyAccount

Some type that is able to be collapsed into an account ID. It is not possible to recreate the original value from the account ID.

Verify

Means of signature verification.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

Balance

Balance is a generic type for the balance of an account.

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Index

Index of a transaction in the chain.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.