Module common_primitives::schema

Structs and traits for the Schema pallet

Structs

SchemaResponse

RPC Response form for a Schema

Enums

ModelType

Types of modeling in which a message payload may be defined

PayloadLocation

Types of payload locations

Traits

SchemaProvider

This allows other pallets to resolve Schema information. With generic SchemaId

SchemaValidator

This allows other Pallets to check validity of schema ids.

Type Definitions

SchemaId

Schema Id is the unique identifier for a Schema