Module common_primitives::msa

Structs and traits for the MSA pallet.

Re-exports

pub use crate::ds::OrderedSetExt;

pub use crate::schema::SchemaId;

Structs

Delegator

A Delegator is a role for an MSA to play. Delegators delegate to Providers.

KeyInfoResponse

RPC Response for getting getting MSA ids

Provider

Provider is the recipient of a delegation. It is a subset of an MSA

ProviderInfo

Struct for the information of the relationship between an MSA and a Provider

ProviderMetadata

This is the metadata associated with a provider. As of now it is just a name, but it will likely be expanded in the future

Constants

EXPIRATION_BLOCK_VALIDITY_GAP

The gap between the current block and a future expiring block allowed when validating signature proofs. 150 blocks at 6 seconds per block would equate to a 15 minute gap.

Traits

AccountProvider

This allows other pallets to resolve MSA information.

Type Definitions

MessageSourceId

Message Source Id or msaId is the unique identifier for Message Source Accounts