Module common_primitives::msa

Structs and traits for the MSA pallet.

Structs

Delegator

A Delegator is a role for an MSA to play. Delegators delegate to Providers.

KeyInfo

KeyInfo holds the information on the relationship between a key and an MSA

KeyInfoResponse

RPC Response form of KeyInfo

Provider

Provider is the recipient of a delegation. It is a subset of an MSA

ProviderInfo

Struct for the information of the relationship between an MSA and a Provider

Traits

AccountProvider

This allows other pallets to resolve MSA information.

Type Definitions

MessageSourceId

Message Source Id or msaId is the unique identifier for Message Source Accounts